+++ src/index.js
@@ -7,6 +7,10 @@ module.exports = function(config = {}) {
       return callback(new Error("git not found on your system."));
     }
 
+    if (config.gitCwd !== undefined && typeof config.gitCwd !== "string") {
+      return callback(new Error("the 'gitCwd' option must be a string."));
+    }
+
     git.stage(file.path, config, error => {
       if (error) {
         return callback(new Error("git add failed."));